Move multisig to dedicated service
