+++ src/app/components/multisig/multisig.component.css
@@ -27,5 +27,5 @@
 
 .data-text-box {
   height: 80px;
-  line-height: 20px;
+  line-height: 25px;
 }
