+++ src/app/components/multisig/multisig.component.html
@@ -12,10 +12,18 @@
             </p>
 
             <h4>Create a Multisig Account</h4>
-            <p>Add 2 or more regular Nano accounts to form a new multisig account.<br>
-            Each account MUST have a known corresponding private key (or seed) or your Nano will be locked.</p>
-            
-            <p> This multisig is "m of m", meaning all keys must be available. Losing one key will invalidate the account.</p>
+            <p>Add 2 or more regular Nano accounts to form a new multisig account.</p>
+            <div>
+              <div class="uk-alert uk-alert-danger nlt-inline-alert">
+                <div class="icon-column">
+                  <div uk-icon="icon: warning; ratio: 2;"></div>
+                </div>
+                <p>
+                  Each account MUST have a known corresponding private key (or seed).<br>
+                  This multisig is "m of m", meaning all keys must be available. <strong>Losing one key will invalidate the account!</strong><br>
+                </p>
+              </div>
+            </div>
             <div class="details-header advanced-options-link" (click)="showAdvancedOptions = !showAdvancedOptions">
               MORE INFO
               <div [class]="['chevron', ( showAdvancedOptions ? 'chevron-up' : 'chevron-down' )]"></div>
@@ -42,6 +50,7 @@
                 </div>
               </div>
             </div>
+
             <div class="uk-form-horizontal">
               <div class="uk-margin">
                 <label class="uk-form-label" for="form-horizontal-text2">Stored Accounts <span uk-icon="icon: info;" uk-tooltip title="Accounts to participate in the multisig account"></span></label>
@@ -52,15 +61,34 @@
                 </div>
               </div>
             </div>
+
+            <div class="uk-form-horizontal">
+              <div class="uk-margin">
+                <label class="uk-form-label" for="form-horizontal-text2">Multisig Account <span uk-icon="icon: info;" uk-tooltip title="Generated account to use with multisig transactions"></span></label>
+                <div class="uk-form-controls">
+                  <div class="uk-inline uk-width-1-1">
+                    <a *ngIf="createdMultisig !== ''" class="uk-form-icon uk-form-icon-flip" ngxClipboard [cbContent]="createdMultisig" (cbOnSuccess)="copied()" uk-icon="icon: copy" title="Copy Account" uk-tooltip></a>
+                    <input [(ngModel)]="createdMultisig" class="uk-input" disabled type="text" placeholder="" autocomplete="off">
+                  </div>
+                </div>
+              </div>
+            </div>
             
             <div class="uk-inline uk-width-1-1">
-              <button class="uk-button uk-button-primary uk-width-1-4@l uk-width-1-1" (click)="generateMultisig()">GENERATE ACCOUNT</button>
+              <div uk-grid>
+                <div class="uk-width-1-3@l uk-width-1-1">
+                  <button class="uk-button uk-button-primary uk-width-1-1" (click)="generateMultisig()">GENERATE MULTISIG</button>
+                </div>
+                <div class="uk-width-1-3@l uk-width-1-1">
+                  <button class="uk-button uk-button-primary uk-width-1-1" (click)="reset()">RESET</button>
+                </div>
+              </div>
             </div>
           </div>
         </div>
         <div class="uk-card uk-card-default" style="margin-top: 20px;">
           <div class="uk-card-header uk-form-horizontal">
-            <h4>Sign Transaction: Prepare</h4>
+            <h4>Create a Multisig Transaction</h4>
             <p>Create a SEND, RECEIVE or CHANGE block. No login is needed.</p>
             
             <div class="uk-form-horizontal">
@@ -76,11 +104,15 @@
               </div>
             </div>
             <div class="uk-inline uk-width-1-1">
-              <button class="uk-button uk-button-primary uk-width-1-4@l uk-width-1-1" (click)="navigateAccount()">CREATE BLOCK</button>
+              <div uk-grid>
+                <div class="uk-width-1-3@l uk-width-1-1">
+                  <button class="uk-button uk-button-primary uk-width-1-1" (click)="navigateAccount()">CREATE BLOCK</button>
+                </div>
+              </div>
             </div>
           </div>
           <div class="uk-card-header uk-form-horizontal">
-            <h4>Sign Transaction: Participant(s)</h4>
+            <h4>Multisignature for Participant(s)</h4>
             <p>Use the data that was shared with you to participate in the signing procedure.</p>
             
             <div class="uk-form-horizontal">
@@ -94,7 +126,11 @@
               </div>
             </div>
             <div class="uk-inline uk-width-1-1">
-              <button class="uk-button uk-button-primary uk-width-1-4@l uk-width-1-1" (click)="navigateBlock(unsignedBlock)">SIGN</button>
+              <div uk-grid>
+                <div class="uk-width-1-3@l uk-width-1-1">
+                  <button class="uk-button uk-button-primary uk-width-1-1" (click)="navigateBlock(unsignedBlock)">SIGN</button>
+                </div>
+              </div>
             </div>
           </div>
         </div>
