+++ src/app/components/sign/sign.component.css
@@ -96,5 +96,5 @@
 
 .data-text-box {
   height: 80px;
-  line-height: 20px;
+  line-height: 25px;
 }
