+++ src/app/components/sign/sign.component.html
@@ -216,7 +216,7 @@
                       <div class="uk-width-1-1">
                         <div class="uk-inline uk-width-1-1" style="height: 40px; margin-bottom: 10px;">
                           <a class="uk-form-icon uk-form-icon-flip" uk-icon="icon: plus" (click)="addMultisigInputData()" uk-tooltip title="Add input data"></a>
-                          <input [(ngModel)]="inputAdd" [disabled]="this.isInvalidStage || this.isInputAddDisabled" maxLength="130" class="uk-input uk-margin-small-bottom {{validInputAdd ? '':'uk-form-danger'}}" id="input-data" type="text" placeholder="{{(this.activeStep - 1) + ':abc...'}}" (ngModelChange)="inputAddChange(inputAdd)" (keyup.enter)="addMultisigInputData()" autocomplete="off">
+                          <input [(ngModel)]="inputAdd" [disabled]="this.isInputAddDisabled" maxLength="130" class="uk-input uk-margin-small-bottom {{validInputAdd ? '':'uk-form-danger'}}" id="input-data" type="text" placeholder="{{(this.activeStep - 1) + ':abc...'}}" (ngModelChange)="inputAddChange(inputAdd)" (keyup.enter)="addMultisigInputData()" autocomplete="off">
                         </div>
                       </div>
                     </div>
@@ -285,7 +285,7 @@
           </div>
           <div class="uk-width-1-2@s">
             <button *ngIf="!confirmingTransaction && shouldSign && signTypeSelected !== signTypes[3]" class="uk-button uk-button-primary uk-width-1-1" (click)="confirmTransaction()">Confirm & Sign</button>
-            <button *ngIf="!confirmingTransaction && shouldSign && signTypeSelected === signTypes[3]" class="uk-button uk-button-primary uk-width-1-1" (click)="startMultisig()" [disabled]="!this.validPrivkey || !this.validParticipants || this.isInvalidStage || (this.activeStep > 1 && !this.isInputAddDisabled)">{{this.activeStep !== 4 ? (this.activeStep === 1 ? 'Start Multi-signing' : 'Step '+ (this.activeStep - 1) + '/3 | Next') : 'Step '+ (this.activeStep - 1) + '/3 | Final Step'}}</button>
+            <button *ngIf="!confirmingTransaction && shouldSign && signTypeSelected === signTypes[3]" class="uk-button uk-button-primary uk-width-1-1" (click)="startMultisig()" [disabled]="!this.validPrivkey || !this.validParticipants || (this.activeStep > 1 && !this.isInputAddDisabled)">{{this.activeStep !== 4 ? (this.activeStep === 1 ? 'Start Multi-signing' : 'Step '+ (this.activeStep - 1) + '/3 | Next') : 'Step '+ (this.activeStep - 1) + '/3 | Final Step'}}</button>
             <button *ngIf="confirmingTransaction && shouldSign" class="uk-button uk-button-primary uk-disabled uk-width-1-1"><span class="uk-margin-right" uk-spinner></span> Loading</button>
             <button *ngIf="!confirmingTransaction && !shouldSign" class="uk-button uk-button-primary uk-width-1-1" (click)="confirmBlock()">Confirm & Process</button>
             <button *ngIf="confirmingTransaction && !shouldSign" class="uk-button uk-button-primary uk-disabled uk-width-1-1"><span class="uk-margin-right" uk-spinner></span> Loading</button>
